# tmanager/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tenant Manager API",
    "description": "API for managing Tenants, Participant Profiles, Cells, and Dataspace Profiles",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/v1alpha1/cells": {
      "get": {
        "tags": [
          "Not implemented yet"
        ],
        "summary": "List Cells",
        "description": "Retrieve all Cells",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1Alpha1Cell"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Cell",
        "description": "Create a new Cell",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1Alpha1NewCell"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Alpha1Cell"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1alpha1/dataspace-profiles": {
      "get": {
        "tags": [
          "Not implemented yet",
          "Not implemented yet"
        ],
        "summary": "List Dataspace Profiles",
        "description": "Retrieve all dataspace profiles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1Alpha1DataspaceProfile"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Dataspace Profile",
        "description": "Create a new Dataspace Profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1Alpha1NewDataspaceProfile"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Alpha1DataspaceProfile"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1alpha1/dataspace-profiles/{id}": {
      "get": {
        "tags": [
          "Not implemented yet"
        ],
        "summary": "Get Dataspace Profile",
        "description": "Retrieve a Dataspace Profile by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Alpha1DataspaceProfile"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1alpha1/dataspace-profiles/{id}/deployments": {
      "post": {
        "summary": "Deploy a Dataspace Profile",
        "description": "Deploy a Dataspace Profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/v1alpha1/tenants": {
      "get": {
        "tags": [
          "Not implemented yet"
        ],
        "summary": "List Tenants",
        "description": "Retrieve all Tenants",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1Alpha1Tenant"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new Tenant",
        "description": "Create a new Tenant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1Alpha1NewTenant"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Alpha1Tenant"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1alpha1/tenants/query": {
      "post": {
        "summary": "Perform a Tenant query",
        "description": "Perform a Tenant query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1Alpha1Tenant"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1alpha1/tenants/{id}": {
      "get": {
        "tags": [
          "Not implemented yet"
        ],
        "summary": "Get Tenant",
        "description": "Retrieve a Tenant by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Alpha1Tenant"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1alpha1/tenants/{id}/participants": {
      "get": {
        "tags": [
          "Not implemented yet"
        ],
        "summary": "List Participant Profiles",
        "description": "Retrieve all Participant Profiles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1Alpha1ParticipantProfile"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Participant Profile",
        "description": "Create a new Participant Profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1Alpha1ParticipantProfile"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Alpha1ParticipantProfile"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1alpha1/tenants/{id}/participants/{participantID}": {
      "delete": {
        "summary": "Dispose Participant Profiles",
        "description": "Dispose a Participant Profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "participantID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Alpha1ParticipantProfile"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ModelQuery": {
        "required": [
          "predicate"
        ],
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "predicate": {
            "type": "string"
          }
        }
      },
      "V1Alpha1Cell": {
        "required": [
          "id",
          "version",
          "state",
          "stateTimestamp"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          },
          "state": {
            "type": "string"
          },
          "stateTimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "V1Alpha1DataspaceDeployment": {
        "required": [
          "id",
          "version",
          "state",
          "stateTimestamp"
        ],
        "type": "object",
        "properties": {
          "cellId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          },
          "state": {
            "type": "string"
          },
          "stateTimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "V1Alpha1DataspaceProfile": {
        "required": [
          "id",
          "version"
        ],
        "type": "object",
        "properties": {
          "artifacts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deployments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/V1Alpha1DataspaceDeployment"
            }
          },
          "id": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "V1Alpha1NewCell": {
        "required": [
          "state",
          "stateTimestamp"
        ],
        "type": "object",
        "properties": {
          "properties": {
            "type": "object",
            "additionalProperties": {}
          },
          "state": {
            "type": "string"
          },
          "stateTimestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "V1Alpha1NewDataspaceProfile": {
        "type": "object",
        "properties": {
          "artifacts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
      "V1Alpha1NewTenant": {
        "type": "object",
        "properties": {
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
      "V1Alpha1ParticipantProfile": {
        "required": [
          "id",
          "version",
          "identifier"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "errorDetail": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "vpas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/V1Alpha1VirtualParticipantAgent"
            }
          }
        }
      },
      "V1Alpha1Tenant": {
        "required": [
          "id",
          "version"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "V1Alpha1VirtualParticipantAgent": {
        "required": [
          "id",
          "version",
          "state",
          "stateTimestamp",
          "type",
          "cell"
        ],
        "type": "object",
        "properties": {
          "cell": {
            "$ref": "#/components/schemas/V1Alpha1Cell"
          },
          "id": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          },
          "state": {
            "type": "string"
          },
          "stateTimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  }
}